chore: consolidate WavelengthForm stories
